Moving cprint to new visualization package
