+++ carla/recourse_methods/catalog/clue/library/clue_ml/AE_models/AE/train.py
@@ -7,6 +7,7 @@ from numpy.random import normal
 from torchvision.utils import make_grid
 
 from carla.recourse_methods.catalog.clue.library.clue_ml.src.utils import *
+from carla.visualisation import cprint
 
 
 def train_VAE(
